UHF-8737: rest of the things in their own preprocesses
